Initial ps4 support + fix's
Added Initial PS4 Support
Re-did the source to actually load rdr as main target and the correct platform depending on source.
Removed un-used files.

Some of the switch left and right based menus/options "should?" now have correct ui button prompts.
I forgot to update them on last source update push.

Initial ps4 support is "as is" atm till I get more time to focus on it.
I've had little time to test this.
The ui button prompts "should?" be fixed from where people were using the switch scripts.
However if it shows a button prompt that is not the correct prompt, please let me know and i'll fix it.
Song player (at least on my end) is not currently working properly on ps4, that or the volume is just way too low. It is being looked into.
Shoutout to DEv_ShOoTz for bringing it to my attention that the menu works on ps4! He has been credited as a Tester for this in the ps4 versions of the menu.