Undead Interior Loader + info
Added the interior loader to the undead scripts.

As is, I only added the catacombs as a undead specific interior to load.
Undead has alot of small misc "props" that can get spawned via interior loading but nothing woth
there salt. Only decent one is the catacombs. On old old scripts on older projects I also made it so you could enable
and disable the undead fire's around the world. However i'm currently at "script breaking point".

What is script breaking point???
Script breaking point is something us rdr1 script modders like to call end of the line. The script can't go any futher without compile corruption.
This is a issue with the compiler we use, once you reach the end of the line the script will either not compile or it will compile corrupted.
There is no known fix for this, even the dev of SC-CL tried to fix this issue for rdr1 and it got us a bit further but in the end the issue always comes
back up.

So what does the future hold for this menu!?
While it's not a path I want to go down it's out of my control on the matter.
All "major" future updates will be pushed to a 2nd script that can be loaded with the script loader.
So far things that will eventually be removed from the main script that will be pushed to the 2nd script (assuming I get that far)
will be the cutscene player, and a object spawner. Cutscenes still need 100% mapped out. I have a object spawner that works currently
but it's very simplistic in nature but i'm having trouble with rotations for spawned objects (i'm not that good with the math for that so I will eventually rely on external help for that). My plan is to start working on the 2nd script soon but it prob won't be pushed to the repo for awhile, at least not till cutscenes are bare minimum fully mapped out.